pyproject.toml
[tool.poetry]
name = "database-api"
version = "0.1.0"
description = "database management system"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
pydantic-settings = "^2.1.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.27"
pyjwt = "^2.8.0"
bcrypt = "^4.1.2"
python-jose = "^3.3.0"
passlib = "^1.7.4"
sqlalchemy-utils = "^0.41.1"
mysql-connector-python = "^8.3.0"
python-dotenv = "^1.0.1"
alembic = "^1.13.1"
requests = "^2.31.0"
pytest = "7.4.2"
pytest-cov = "4.1.0"
httpcore = "^1.0.4"
httpx = "^0.27.0"
pyyaml = "^6.0.1"
pylint = "^3.1.0"
coverage = "^6.1.1"
coveralls = "^3.3.1"
mypkg = "^1.0.0"
starlette= "0.36.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
